Retrieve dataframe response from Geoclient for addresses

This function takes BBLs (borough-block-lot) and returns Geoclient response in a tibble. The BBLs can be provided either in a vector as a named argument or as a dataframe and column name of the BBL field. The Geoclient API's app ID and key can either be provided directly as arguments, or you can first use geoclient_api_keys() to add them to your .Renviron file so they can be called securely without being stored in your code. You can acquire your Geoclient app ID and Key by first registering with the https://developer.cityofnewyork.us/user/register?destination=api(NYC Developer Portal) at, then https://developer.cityofnewyork.us/create/project(create a new project), selecting "Geoclient v1" from available APIs.

geoclient_address(df = NULL, number, street, borough = NULL, zip = NULL,
  id = NULL, key = NULL)

Arguments

df
Dataframe that contains a column of BBLs. Defaults to NULL and bbl is taked as a vector.

id
The API app ID provided to you from the NYC Developer Portal formated in quotes. Defaults to NULL and your key is accessed from your .Renviron.

key
The API app key provided to you from the NYC Developer Portal formated in quotes. Defaults to NULL and your key is accessed from your .Renviron.

bbl
Either a vector of BBLs (numeric or character is accepted), or a bare column name of the bbl field if a dataframe is provided.

Examples

# NOT RUN {
geoclient_api_keys("1a2b3c4", "9d8f7b6wh4jfgud67s89jfyw68vj38fh", install = TRUE)
# First time, reload your environment so you can use the keys without restarting R.
readRenviron("~/.Renviron")

geoclient_address(
  number = "139",
  street = "MacDougal St",
  borough = "manhattan",
  zip = "11231"
)

library(tibble)
library(dplyr)

addr <- tribble(
  ~num,  ~st,                ~boro,         ~zip,
  "139", "MacDougal St",     "manhattan",   "11231",
  "295", "Lafayette street", NA_character_, "10012-2722",
  "40",  "WASHINGTON SQ S",  "MN",          NA_character_
)

geoclient_address(addr, num, st, boro, zip)

addr %>%
  mutate(
    bbl = geoclient_address(number = num, street = st, borough = boro, zip = zip)[["bbl"]]
  )
# }