Retrieve Geoclient Response for Blockfaces as a Dataframe

This function takes components of addresses and returns the Geoclient response as a tibble. The house number, street name, and one of either borough or Zip code are required. The address components can be provided either in separate vectors as named arguments or with a dataframe and column names containing each component. The Geoclient API's app ID and key can either be provided directly as arguments, or you can first use geoclient_api_key() to add them to your .Renviron file so they can be called securely without being stored in your code.

geoclient_intersection(df = NULL, cross_street_1, cross_street_2, borough,
  cross_street_2_borough = NULL, compass_direction = NULL, id = NULL,
  key = NULL, rate_limit = TRUE, cap_daily_requests = TRUE)

Arguments

df
Dataframe that contains a column of BBLs. Defaults to NULL and bbl is taken as a vector.

cross_street_1
The first cross street of the blockface, as either a vector of street names, or a bare column name of the field if a dataframe is provided.

cross_street_2
The second cross street of the blockface, as either a vector of street names, or a bare column name of the field if a dataframe is provided.

borough
The name of the borough of the the first cross street, or the entire intersection if cross_street_2_borough is not provided. The argument can be provided as either a vector or a bare column name of the borough field if a dataframe is provided.

cross_street_2_borough
Optionally, the name of the borough of the second cross street if it differs from the first cross street. The argument can be provided as either a vector or a bare column name of the borough field if a dataframe is provided.

compass_direction
Optionally, the direction indicating which intersection is desired when the two street intersect in multiple locations. The argument can be provided as either a vector or a bare column name of the field if a dataframe is provided. The valid values of are "N", "E", "S", or "W"

id
The API app ID provided to you from the NYC Developer Portal formated in quotes. Defaults to NULL and your key is accessed from your .Renviron.

key
The API app key provided to you from the NYC Developer Portal formated in quotes. Defaults to NULL and your key is accessed from your .Renviron.

rate_limit
Whether you would like to limit the rate of API requests in adherence to Geoclient's Service Usage Guidelines. See ?geoclient for more information.

cap_daily_requests
Whether you would like to cap the daily number of API requests in adherence to Geoclient's Service Usage Guidelines. See ?geoclient for more information.

Details

For more details see the Geoclient Documentation's guide to making address requests, interpreting the Geosupport return codes, the data returned by geoclient_address, and a complete data dictionary for all possible data elements returned by any geoclient function.

Examples

# NOT RUN {
geoclient_api_key("1a2b3c4", "9d8f7b6wh4jfgud67s89jfyw68vj38fh")

geoclient_intersection(
  cross_street_1 = "macdougal st",
  cross_street_2 = "w 3rd st",
  borough = "mn"
)

library(tibble)
library(dplyr)

df <- tribble(
  ~cross1,         ~cross2,       ~boro,
  "macdougal st",  "w 3rd st",    "mn",
  "Lexington Ave", "125th Steet", "Manhattan"
)

geoclient_intersection(df, cross1, cross2, boro)

df %>%
  mutate(
    lion_node_number = geoclient_intersection(
      cross_street_1 = cross1,
      cross_street_2 = cross2,
      borough = boro
      )[["lionNodeNumber"]]
  )
# }