Retrieve Geoclient Response for Any Location Type as a Dataframe

This function takes any of the 6 location types as a single input and returns the Geoclient response as a tibble. The locations are provided either in a single vector as a named argument or with a dataframe and column name of the location field. The Geoclient API's app ID and key can either be provided directly as arguments, or you can first use geoclient_api_keys() to add them to your .Renviron file so they can be called securely without being stored in your code.

geoclient_search(df = NULL, location, id = NULL, key = NULL,
  rate_limit = TRUE, cap_daily_requests = TRUE)

Arguments

df
Dataframe that contains a column of BBLs. Defaults to NULL and bbl is taken as a vector.

location
Any of the 6 locations types from other functions: Address, BBL, BIN, Blockface, Intersection, or Place. The argument can be either a single vector of locations, or a bare column name of the location field if a dataframe is provided.

id
The API app ID provided to you from the NYC Developer Portal formated in quotes. Defaults to NULL and your key is accessed from your .Renviron.

key
The API app key provided to you from the NYC Developer Portal formated in quotes. Defaults to NULL and your key is accessed from your .Renviron.

rate_limit
Whether you would like to limit the rate of API requests in adherence to Geoclient's Service Usage Guidelines. See ?geoclient for more information.

cap_daily_requests
Whether you would like to cap the daily number of API requests in adherence to Geoclient's Service Usage Guidelines. See ?geoclient for more information.

Details

Geoclient is quite flexible with the format of the address when provided as a single input for this function, however the results may be slower than the more restrictive geoclient_address() because Geoclient make have to make multiple requests to narrow in on the correct location.

For more details see the Geoclient Documentation's guide to single-input requests, interpreting the Geosupport return codes, and a complete data dictionary for all possible data elements returned by any geoclient function.

Examples

# NOT RUN {
geoclient_api_keys("1a2b3c4", "9d8f7b6wh4jfgud67s89jfyw68vj38fh")

geoclient_search(location = 1005430053)
geoclient_search(location = "139 macdougal st mn")
geoclient_search(location = c("1008760", "1007941"))

library(dplyr)

df <- tibble(loc = c("1005430053", "139 MacDougal Street, New York, 10012", "1008760"))

geoclient_search(df, loc)
# }